[FEAT] implement PeepCheckButton widget
